+++ Engine.h
@@ -14,8 +14,8 @@ public:
     void checkUp(SEARCHINFO*);
     void pickNextMove(int, Movelist*);
     void clearForSearch(Board*, SEARCHINFO*);
-    double alphaBeta(double, double, int, Board*, SEARCHINFO*);
-    double quiescence(double, double, Board*, SEARCHINFO*);
+    int alphaBeta(int, int, int, Board*, SEARCHINFO*);
+    int quiescence(int, int, Board*, SEARCHINFO*);
     void searchPosition(Board*, SEARCHINFO*, bool);
 };
